Make Node contain its Rc<RefCell>>
The primary motivation for this was
so add_child could be responsible for
setting the child's parent. Parent
should only ever be set by calling
add_child, thereby avoiding parent
and child becoming out of sync.

The extra benefit of this change is
Rc<RefCell<NodeData>> is now an
implementation detail of Node, so
Rc and RefCell don't need to be
used in the main function
at all.

NodeData is a class that should
always remain private to the node
module.